:truck: Move the InMemoryFS.kt to test-utils module
