+++ koog-cli/src/commonTest/kotlin/ai/koog/agents/ext/tool/edit/EditFileToolMultilineContentTest.kt
@@ -2,9 +2,9 @@ package ai.koog.agents.ext.tool.edit
 
 import ai.koog.agents.core.tools.DirectToolCallsEnabler
 import ai.koog.agents.core.tools.annotations.InternalAgentToolsApi
-import ai.koog.agents.ext.tool.edit.testutil.InMemoryFS
 import ai.koog.rag.base.files.readText
 import ai.koog.rag.base.files.writeText
+import ai.koog.test.utils.InMemoryFS
 import kotlinx.coroutines.test.runTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
