+++ test-utils/src/commonMain/kotlin/ai/koog/test/utils/InMemoryFS.kt
@@ -1,4 +1,4 @@
-package ai.koog.agents.ext.tool.edit.testutil
+package ai.koog.test.utils
 
 import ai.koog.rag.base.files.FileMetadata
 import ai.koog.rag.base.files.FileSystemProvider
@@ -9,7 +9,7 @@ import kotlinx.io.Source
  * Minimal in-memory filesystem for tests. Implements ReadWrite<String> and stores text as ByteArray.
  * Shared across integration tests to avoid duplication.
  */
-class InMemoryFS : FileSystemProvider.ReadWrite<String> {
+public class InMemoryFS : FileSystemProvider.ReadWrite<String> {
     private val files = mutableMapOf<String, ByteArray>()
     private val directories = mutableSetOf<String>()
